reorganized

+++ utils/download.sh
@@ -1,10 +1,10 @@
 # This script basically downloads the entire dataset and converts them to required binary format used by the program.
+#!/bin/bash 
 
-if [[ -z "${PROJREPO}"]]; then 
+if [[ -z "${PROJREPO}" ]]; then 
        echo "ERROR: Set PROJREPO enviornment variable"	
        exit 0
-if
-
+fi
 
 echo "Downloading Categories"
 wget https://graphchallenge.s3.amazonaws.com/synthetic/sparsechallenge_2019/dnn/neuron1024-l120-categories.tsv 
@@ -46,7 +46,7 @@ gzip -d sparse-images-16384.tsv.gz
 gzip -d sparse-images-65536.tsv.gz
 
 echo "Converting Sparse Images to Binary" 
-g++ ../convert_sparsesimages.cpp -o convertimages
+g++ $PROJREPO/utils/convert_sparseimages.cpp  -o convertimages
 ./convertimages 1024
 ./convertimages 4096
 ./convertimages 16384
@@ -60,13 +60,23 @@ tar vxfz neuron65536.tar.gz
 
 
 echo "Converting Weights to Binary"
-g++ ../convert_weights.cpp -o convertneuron 
+g++ $PROJREPO/utils/convert_weights.cpp -o convertneuron 
 ./convertneuron 1024
 ./convertneuron 4096
 ./convertneuron 16384
 ./convertneuron 65536
 
-echo "Cleanup not needed files."
-rm -rf sparse-image*.tsv*
-rm -rf neuron1024 neuron4096 neuron16384 neuron65536
-rm -rf neuron1024.tar.gz neuron4096.tar.gz neuron16384.tar.gz neuron65536.tar.gz
+
+echo "Is the dataset downloaded successfully and uncompressed? Enter [Y/N]:"
+read success
+
+if [[ "$success" == "Y" ]]; then 
+	echo "Cleanup not needed files."
+	rm -rf sparse-image*.tsv*
+	rm -rf neuron1024 neuron4096 neuron16384 neuron65536
+	rm -rf neuron1024.tar.gz neuron4096.tar.gz neuron16384.tar.gz neuron65536.tar.gz
+else 
+	echo "Skipping cleanup!"
+fi
+
+echo "Done"
